Migrate the last of the non-rule tests to jest
