+++ .travis.yml
@@ -5,22 +5,21 @@ cache:
     - node_modules
 matrix:
   include:
-    - node_js: '6'
-      script: npm run lint && npm run flow
-      env: CI=lint+flow
     - node_js: '7'
-      script: npm run jest-ci && npm run test
-      env: CI=test 7
+      script: npm run pretest
+      env: CI=pretest
+    - node_js: '7'
+      script: npm run test:jest --- --runInBand && npm run test:tape
+      env: CI=tests 7
     - node_js: '6'
-      script: npm run jest-ci && npm run test
-      env: CI=test 6
+      script: npm run test:jest -- --runInBand --testPathPattern /lib/ && npm run test:tape
+      env: CI=tests 6
     - node_js: '4'
-      script: npm run jest-ci && npm run test
-      env: CI=test 4
+      script: npm run test:jest -- --runInBand --testPathPattern /lib/ && npm run test:tape
+      env: CI=tests 4
 before_install:
   - npm prune
   - npm update
-after_success: npm run coveralls
 notifications:
   webhooks:
     - https://webhooks.gitter.im/e/d763493612da45967361
