+++ appveyor.yml
@@ -3,7 +3,6 @@
 environment:
   matrix:
     - nodejs_version: 7
-    - nodejs_version: 6
 
 version: "{build}"
 build: off
@@ -20,4 +19,4 @@ install:
 test_script:
   - node --version
   - npm --version
-  - cmd: "npm test"
+  - cmd: "npm run test:jest -- --runInBand --testPathPattern \\lib && npm run test:tape"
