+++ lib/__tests__/disableRanges-integration.test.js
@@ -4,7 +4,6 @@ const blockNoEmpty = require("../rules/block-no-empty")
 const maxLineLength = require("../rules/max-line-length")
 const selectorCombinatorSpaceBefore = require("../rules/selector-combinator-space-before")
 const stringQuotes = require("../rules/string-quotes")
-const testRule = require("../testUtils/testRule.js")
 
 // disabling all rules
 testRule(blockNoEmpty, {
